+++ packages/api/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @gpc-cli/api
 
+## 1.0.21
+
+### Patch Changes
+
+- Resumable uploads, Google best practices compliance, bug fixes
+
 ## 1.0.20
 
 ### Patch Changes